move toolbar logic to separate file
