+++ CodeEdit/Features/Documents/Controllers/CodeEditWindowController+Toolbar.swift
@@ -1,0 +1,127 @@
+//
+//  CodeEditWindowController+Toolbar.swift
+//  CodeEdit
+//
+//  Created by Daniel Zhu on 5/10/24.
+//
+
+import AppKit
+import SwiftUI
+
+extension CodeEditWindowController {
+    internal func setupToolbar() {
+        let toolbar = NSToolbar(identifier: UUID().uuidString)
+        toolbar.delegate = self
+        toolbar.displayMode = .labelOnly
+        toolbar.showsBaselineSeparator = false
+        self.window?.titleVisibility = toolbarCollapsed ? .visible : .hidden
+        self.window?.toolbarStyle = .unifiedCompact
+        if Settings[\.general].tabBarStyle == .native {
+            // Set titlebar background as transparent by default in order to
+            // style the toolbar background in native tab bar style.
+            self.window?.titlebarSeparatorStyle = .none
+        } else {
+            // In Xcode tab bar style, we use default toolbar background with
+            // line separator.
+            self.window?.titlebarSeparatorStyle = .automatic
+        }
+        self.window?.toolbar = toolbar
+    }
+
+    func toolbarDefaultItemIdentifiers(_ toolbar: NSToolbar) -> [NSToolbarItem.Identifier] {
+        [
+            .toggleFirstSidebarItem,
+            .sidebarTrackingSeparator,
+            .branchPicker,
+            .flexibleSpace,
+            .itemListTrackingSeparator,
+            .flexibleSpace,
+            .toggleLastSidebarItem
+        ]
+    }
+
+    func toolbarAllowedItemIdentifiers(_ toolbar: NSToolbar) -> [NSToolbarItem.Identifier] {
+        [
+            .toggleFirstSidebarItem,
+            .sidebarTrackingSeparator,
+            .flexibleSpace,
+            .itemListTrackingSeparator,
+            .toggleLastSidebarItem,
+            .branchPicker
+        ]
+    }
+
+    func toggleToolbar() {
+        toolbarCollapsed.toggle()
+        updateToolbarVisibility()
+    }
+
+    private func updateToolbarVisibility() {
+        if toolbarCollapsed {
+            window?.titleVisibility = .visible
+            window?.title = workspace?.workspaceFileManager?.folderUrl.lastPathComponent ?? "Empty"
+            window?.toolbar = nil
+        } else {
+            window?.titleVisibility = .hidden
+            setupToolbar()
+        }
+    }
+
+    func toolbar(
+        _ toolbar: NSToolbar,
+        itemForItemIdentifier itemIdentifier: NSToolbarItem.Identifier,
+        willBeInsertedIntoToolbar flag: Bool
+    ) -> NSToolbarItem? {
+        switch itemIdentifier {
+        case .itemListTrackingSeparator:
+            guard let splitViewController else { return nil }
+
+            return NSTrackingSeparatorToolbarItem(
+                identifier: .itemListTrackingSeparator,
+                splitView: splitViewController.splitView,
+                dividerIndex: 1
+            )
+        case .toggleFirstSidebarItem:
+            let toolbarItem = NSToolbarItem(itemIdentifier: NSToolbarItem.Identifier.toggleFirstSidebarItem)
+            toolbarItem.label = "Navigator Sidebar"
+            toolbarItem.paletteLabel = " Navigator Sidebar"
+            toolbarItem.toolTip = "Hide or show the Navigator"
+            toolbarItem.isBordered = true
+            toolbarItem.target = self
+            toolbarItem.action = #selector(self.toggleFirstPanel)
+            toolbarItem.image = NSImage(
+                systemSymbolName: "sidebar.leading",
+                accessibilityDescription: nil
+            )?.withSymbolConfiguration(.init(scale: .large))
+
+            return toolbarItem
+        case .toggleLastSidebarItem:
+            let toolbarItem = NSToolbarItem(itemIdentifier: NSToolbarItem.Identifier.toggleLastSidebarItem)
+            toolbarItem.label = "Inspector Sidebar"
+            toolbarItem.paletteLabel = "Inspector Sidebar"
+            toolbarItem.toolTip = "Hide or show the Inspectors"
+            toolbarItem.isBordered = true
+            toolbarItem.target = self
+            toolbarItem.action = #selector(self.toggleLastPanel)
+            toolbarItem.image = NSImage(
+                systemSymbolName: "sidebar.trailing",
+                accessibilityDescription: nil
+            )?.withSymbolConfiguration(.init(scale: .large))
+
+            return toolbarItem
+        case .branchPicker:
+            let toolbarItem = NSToolbarItem(itemIdentifier: .branchPicker)
+            let view = NSHostingView(
+                rootView: ToolbarBranchPicker(
+                    workspaceFileManager: workspace?.workspaceFileManager
+                )
+            )
+            toolbarItem.view = view
+
+            return toolbarItem
+
+        default:
+            return NSToolbarItem(itemIdentifier: itemIdentifier)
+        }
+    }
+}
